requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cryptography==46.0.3
cx_Oracle==8.3.0
cycler==0.12.1
defusedxml==0.7.1
Django==6.0.1
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
fonttools==4.61.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
oauthlib==3.3.1
oracledb==3.4.1
packaging==25.0
pillow==12.1.0
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python3-openid==3.2.0
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
